skills/fig2data/requirements.txt
# Requirements for Fig2Data skill
# Install with: pip install -r requirements.txt

# Core dependencies
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# OCR - Primary text extraction
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
scikit-image>=0.21.0
imageio>=2.28.0

# Plot digitization (optional - requires interaction)
plotdigitizer>=0.3.0

# Visualization (optional - for results)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# PDF to image conversion (optional)
pdf2image>=1.16.0

# For table extraction (optional - not ideal for charts)
# img2table>=0.0.12
# pytesseract>=0.3.10
